requirements.txt
# Core dependencies
weaviate-client>=4.4.0
google-generativeai>=0.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
sentence-transformers>=3.0.0

# Fix protobuf compatibility warnings
protobuf>=5.29.0,<6.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

# Database
pymongo>=4.6.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 